etl/requirements.txt
python-dotenv==0.15.0
psycopg2==2.8.6